refactor: Moved TopBar and NoNetwork moved to separate files
