+++ app/src/main/java/com/vivek/githubapisample/home/presentation/HomePage.kt
@@ -3,22 +3,17 @@ package com.vivek.githubapisample.home.presentation
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.slideInVertically
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
@@ -26,8 +21,6 @@ import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -36,7 +29,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
@@ -48,6 +40,8 @@ import androidx.paging.compose.collectAsLazyPagingItems
 import com.vivek.githubapisample.R
 import com.vivek.githubapisample.common.data.AppResult
 import com.vivek.githubapisample.common.presentation.ValueCallback
+import com.vivek.githubapisample.common.presentation.component.NoNetwork
+import com.vivek.githubapisample.common.presentation.component.TopBar
 import com.vivek.githubapisample.common.presentation.toDisplayString
 import com.vivek.githubapisample.repo.data.Repo
 import com.vivek.githubapisample.repo.presentation.RepoListItem
@@ -126,7 +120,7 @@ fun HomePage(
 
     Scaffold(modifier = modifier,
         snackbarHost = { SnackbarHost(hostState) },
-        topBar = { TopBar(modifier = modifier) }
+        topBar = { TopBar(title = R.string.take_home, modifier = modifier) }
     ) { contentPadding ->
         Column(
             modifier = Modifier
@@ -203,49 +197,6 @@ fun HomePage(
 }
 
 /**
- * A composable function that creates a top app bar.
- *
- * @param modifier The modifier to apply to the top app bar.
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TopBar(modifier: Modifier) {
-    TopAppBar(
-        modifier = modifier.statusBarsPadding(),
-        title = {
-            Text(text = stringResource(id = R.string.take_home))
-        },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = MaterialTheme.colorScheme.primary,
-            titleContentColor = MaterialTheme.colorScheme.onPrimary
-        )
-    )
-}
-
-/**
- * A composable function that displays a message when there is no network connection.
- *
- * @param modifier The modifier to apply to the composable.
- */
-@Composable
-fun NoNetwork(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(Color.Red)
-            .padding(
-                horizontal = MaterialTheme.padding.medium,
-                vertical = MaterialTheme.padding.small
-            )
-    ) {
-        Text(
-            text = stringResource(R.string.no_internet_available),
-            style = MaterialTheme.typography.bodyLarge.copy(MaterialTheme.colorScheme.onPrimary)
-        )
-    }
-}
-
-/**
  * A composable function that renders a text field for entering a user ID and a button for searching.
  *
  * @param username The current user ID.
@@ -293,14 +244,6 @@ fun UserNameField(
 fun UserNameFieldPreview() {
     GitHubApiSampleTheme {
         UserNameField("1", onUsernameChange = {}, onSearchClick = {})
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun NoInternetPreview() {
-    GitHubApiSampleTheme {
-        NoNetwork()
     }
 }
